feat(order-service): update-type-in-dto-request-order-item

+++ order-service/src/main/java/com/ecommerce/order_service/dto/request/OrderItemRequest.java
@@ -23,7 +23,7 @@ public class OrderItemRequest {
     @Min(value = 0, message = "Quantity must greater than zero")
     int quantity;
     @Min(value = 0, message = "Price must greater than zero")
-    BigDecimal price;
+    int price;
     @Min(value = 0, message = "subTotal must greater than zero")
-    BigDecimal subTotal;
+    int subTotal;
 }
